Zphisher is an advanced phishing tool for penetration testing and social engineering attacks. Developed by ByteHackedits, it offers customizable templates, URL spoofing, automated email/SMS sending, and real-time tracking. For educational purposes only; unauthorized use is strictly prohibited
Zphisher: A Powerful Phishing Tool

Zphisher is a advanced phishing tool designed to aid in penetration testing and social engineering attacks. Developed by ByteHackedits, this tool provides a simple and effective way to create and manage phishing campaigns.

Key Features:

- Easy-to-use interface for creating and customizing phishing pages
- Support for multiple phishing templates and themes
- Advanced URL and domain spoofing capabilities
- Automated email and SMS sending for phishing attacks
- Real-time tracking and logging of phishing attempts

Disclaimer:

Zphisher is intended for educational and testing purposes only. It should not be used for malicious or illegal activities. Use of this tool for unauthorized phishing attacks is strictly prohibited and may result in severe consequences.

Credit:

Zphisher was developed by ByteHackedits, a renowned cybersecurity expert and developer.